Removing the 'Vnext' suffix from MSFDrawer and associated types.
